+++ src/onegov.org/onegov/org/app.py
@@ -36,7 +36,7 @@ class OrgApp(Framework, LibresIntegration, ElasticsearchApp, MapboxApp,
     userlinks = directive(directives.UserlinkAction)
 
     #: the version of this application (do not change manually!)
-    version = '0.21.0'
+    version = '0.22.0'
 
     def is_allowed_application_id(self, application_id):
         """ Stops onegov.server from ever passing the request to the org